moved module layouter to presentation layer
