+++ module-queue/src/main/java/ch/hsr/adv/ui/queue/presentation/QueueLayouter.java
@@ -1,4 +1,4 @@
-package ch.hsr.adv.ui.queue.logic;
+package ch.hsr.adv.ui.queue.presentation;
 
 import ch.hsr.adv.commons.core.logic.domain.Module;
 import ch.hsr.adv.commons.core.logic.domain.ModuleGroup;
