+++ module-stack/src/test/java/ch/hsr/adv/ui/stack/presentation/StackLayouterTest.java
@@ -1,11 +1,9 @@
-package ch.hsr.adv.ui.graph.logic;
+package ch.hsr.adv.ui.stack.presentation;
 
 import ch.hsr.adv.commons.core.logic.domain.ModuleGroup;
 import ch.hsr.adv.ui.core.access.FileDatastoreAccess;
 import ch.hsr.adv.ui.core.logic.exceptions.ADVParseException;
-import ch.hsr.adv.ui.core.presentation.widgets.CurvedLabeledEdge;
-import ch.hsr.adv.ui.core.presentation.widgets.LabeledNode;
-import ch.hsr.adv.ui.core.presentation.widgets.SelfReferenceEdge;
+import ch.hsr.adv.ui.stack.logic.StackParser;
 import com.google.gson.Gson;
 import com.google.gson.JsonElement;
 import com.google.inject.Inject;
@@ -13,6 +11,7 @@ import javafx.collections.ObservableList;
 import javafx.scene.Group;
 import javafx.scene.Node;
 import javafx.scene.layout.Pane;
+import javafx.scene.layout.VBox;
 import org.jukito.JukitoRunner;
 import org.junit.Before;
 import org.junit.Test;
@@ -22,20 +21,18 @@ import org.testfx.api.FxToolkit;
 import java.io.File;
 import java.io.IOException;
 import java.net.URL;
-import java.util.List;
 import java.util.concurrent.TimeoutException;
-import java.util.stream.Collectors;
 
 import static org.junit.Assert.assertEquals;
 
 @RunWith(JukitoRunner.class)
-public class GraphLayouterTest {
+public class StackLayouterTest {
     @Inject
     private FileDatastoreAccess reader;
     @Inject
-    private GraphParser testParser;
+    private StackParser testParser;
     @Inject
-    private GraphLayouter sut;
+    private StackLayouter sut;
 
     private ModuleGroup moduleGroup;
 
@@ -56,28 +53,10 @@ public class GraphLayouterTest {
         Pane actual = sut.layout(moduleGroup, null);
 
         // THEN
-        // layouter build 10 elements in total
         ObservableList<Node> children = actual.getChildren();
         Group group = (Group) children.get(0);
-        int graphElementCount = group.getChildren().size();
-        assertEquals(10, graphElementCount);
-
-        // layouter build 5 nodes
-        List<Node> graphNodes = group.getChildren().stream()
-                .filter(e -> e instanceof LabeledNode)
-                .collect(Collectors.toList());
-        assertEquals(5, graphNodes.size());
-
-        // layouter build 1 self reference edge
-        List<Node> graphEdgesSelf = group.getChildren().stream()
-                .filter(e -> e instanceof SelfReferenceEdge)
-                .collect(Collectors.toList());
-        assertEquals(1, graphEdgesSelf.size());
-
-        // layouter build 1 curved edge
-        List<Node> graphEdgesCurved = group.getChildren().stream()
-                .filter(e -> e instanceof CurvedLabeledEdge)
-                .collect(Collectors.toList());
-        assertEquals(2, graphEdgesCurved.size());
+        VBox vBox = (VBox) group.getChildren().get(0);
+        int arrayElementCount = vBox.getChildren().size();
+        assertEquals(2, arrayElementCount);
     }
 }
